+++ Source/Core/VideoBackends/OGL/FramebufferManager.cpp
@@ -394,13 +394,13 @@ XFBSource::~XFBSource()
 }
 
 
-void XFBSource::Draw(const MathUtil::Rectangle<float> &sourcerc,
+void XFBSource::Draw(const MathUtil::Rectangle<int> &sourcerc,
 		const MathUtil::Rectangle<float> &drawrc) const
 {
 	// Texture map xfbSource->texture onto the main buffer
 	glFramebufferTexture2D(GL_READ_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, texture, 0);
 	glBlitFramebuffer(sourcerc.left, sourcerc.bottom, sourcerc.right, sourcerc.top,
-		drawrc.left, drawrc.bottom, drawrc.right, drawrc.top,
+		(int)drawrc.left, (int)drawrc.bottom, (int)drawrc.right, (int)drawrc.top,
 		GL_COLOR_BUFFER_BIT, GL_LINEAR);
 
 	GL_REPORT_ERRORD();
